requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
gunicorn==23.0.0